No longer triggering multiple fetchs
Modified the LLFace.cpp/.h to have new mInCameraVirtualSize and accessor methods.
As well, modified the LLFace::getTextureVirutalSize to have 2 inserts where the mInCameraVirtualSize is captured.

This value represents the actual in camera pixel area with texels factored in.

Static flag is used to indicate that the current camera is main viewer camera as well as that auto scale features is active.